requirements.txt
django>=5.0
djangorestframework
django-cors-headers
gunicorn
whitenoise
python-dotenv
dj-database-url
psycopg2-binary

mailjet-rest
